Complete the separation of the IR compiler frontend from the "Jit"
