+++ Core/Core.vcxproj.filters
@@ -670,6 +670,9 @@
     <ClCompile Include="MIPS\IR\IRInterpreter.cpp">
       <Filter>MIPS\IR</Filter>
     </ClCompile>
+    <ClCompile Include="MIPS\IR\IRFrontend.cpp">
+      <Filter>MIPS\IR</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="ELF\ElfReader.h">
@@ -1228,6 +1231,9 @@
       <Filter>MIPS\IR</Filter>
     </ClInclude>
     <ClInclude Include="MIPS\IR\IRInterpreter.h">
+      <Filter>MIPS\IR</Filter>
+    </ClInclude>
+    <ClInclude Include="MIPS\IR\IRFrontend.h">
       <Filter>MIPS\IR</Filter>
     </ClInclude>
   </ItemGroup>
