+++ Core/MIPS/IR/IRJit.h
@@ -20,11 +20,11 @@
 #include <cstring>
 
 #include "Common/CPUDetect.h"
-#include "Core/MIPS/JitCommon/JitState.h"
 #include "Core/MIPS/JitCommon/JitBlockCache.h"
 #include "Core/MIPS/JitCommon/JitCommon.h"
 #include "Core/MIPS/IR/IRRegCache.h"
 #include "Core/MIPS/IR/IRInst.h"
+#include "Core/MIPS/IR/IRFrontend.h"
 #include "Core/MIPS/MIPSVFPUUtils.h"
 
 #ifndef offsetof
@@ -99,137 +99,6 @@ public:
 private:
 	int size_;
 	std::vector<IRBlock> blocks_;
-};
-
-class IRFrontend : public MIPSFrontendInterface {
-public:
-	IRFrontend(bool startDefaultPrefix);
-	void Comp_Generic(MIPSOpcode op) override;
-
-	void Comp_RunBlock(MIPSOpcode op) override;
-	void Comp_ReplacementFunc(MIPSOpcode op) override;
-
-	// Ops
-	void Comp_ITypeMem(MIPSOpcode op) override;
-	void Comp_Cache(MIPSOpcode op) override;
-
-	void Comp_RelBranch(MIPSOpcode op) override;
-	void Comp_RelBranchRI(MIPSOpcode op) override;
-	void Comp_FPUBranch(MIPSOpcode op) override;
-	void Comp_FPULS(MIPSOpcode op) override;
-	void Comp_FPUComp(MIPSOpcode op) override;
-	void Comp_Jump(MIPSOpcode op) override;
-	void Comp_JumpReg(MIPSOpcode op) override;
-	void Comp_Syscall(MIPSOpcode op) override;
-	void Comp_Break(MIPSOpcode op) override;
-
-	void Comp_IType(MIPSOpcode op) override;
-	void Comp_RType2(MIPSOpcode op) override;
-	void Comp_RType3(MIPSOpcode op) override;
-	void Comp_ShiftType(MIPSOpcode op) override;
-	void Comp_Allegrex(MIPSOpcode op) override;
-	void Comp_Allegrex2(MIPSOpcode op) override;
-	void Comp_VBranch(MIPSOpcode op) override;
-	void Comp_MulDivType(MIPSOpcode op) override;
-	void Comp_Special3(MIPSOpcode op) override;
-
-	void Comp_FPU3op(MIPSOpcode op) override;
-	void Comp_FPU2op(MIPSOpcode op) override;
-	void Comp_mxc1(MIPSOpcode op) override;
-
-	void Comp_DoNothing(MIPSOpcode op) override;
-
-	void Comp_SV(MIPSOpcode op) override;
-	void Comp_SVQ(MIPSOpcode op) override;
-	void Comp_VPFX(MIPSOpcode op) override;
-	void Comp_VVectorInit(MIPSOpcode op) override;
-	void Comp_VMatrixInit(MIPSOpcode op) override;
-	void Comp_VDot(MIPSOpcode op) override;
-	void Comp_VecDo3(MIPSOpcode op) override;
-	void Comp_VV2Op(MIPSOpcode op) override;
-	void Comp_Mftv(MIPSOpcode op) override;
-	void Comp_Vmfvc(MIPSOpcode op) override;
-	void Comp_Vmtvc(MIPSOpcode op) override;
-	void Comp_Vmmov(MIPSOpcode op) override;
-	void Comp_VScl(MIPSOpcode op) override;
-	void Comp_Vmmul(MIPSOpcode op) override;
-	void Comp_Vmscl(MIPSOpcode op) override;
-	void Comp_Vtfm(MIPSOpcode op) override;
-	void Comp_VHdp(MIPSOpcode op) override;
-	void Comp_VCrs(MIPSOpcode op) override;
-	void Comp_VDet(MIPSOpcode op) override;
-	void Comp_Vi2x(MIPSOpcode op) override;
-	void Comp_Vx2i(MIPSOpcode op) override;
-	void Comp_Vf2i(MIPSOpcode op) override;
-	void Comp_Vi2f(MIPSOpcode op) override;
-	void Comp_Vh2f(MIPSOpcode op) override;
-	void Comp_Vcst(MIPSOpcode op) override;
-	void Comp_Vhoriz(MIPSOpcode op) override;
-	void Comp_VRot(MIPSOpcode op) override;
-	void Comp_VIdt(MIPSOpcode op) override;
-	void Comp_Vcmp(MIPSOpcode op) override;
-	void Comp_Vcmov(MIPSOpcode op) override;
-	void Comp_Viim(MIPSOpcode op) override;
-	void Comp_Vfim(MIPSOpcode op) override;
-	void Comp_VCrossQuat(MIPSOpcode op) override;
-	void Comp_Vsgn(MIPSOpcode op) override;
-	void Comp_Vocp(MIPSOpcode op) override;
-	void Comp_ColorConv(MIPSOpcode op) override;
-	void Comp_Vbfy(MIPSOpcode op) override;
-
-	int Replace_fabsf();
-	void DoState(PointerWrap &p);
-	bool CheckRounding();  // returns true if we need a do-over
-	void DoJit(u32 em_address, IRBlock *b);
-
-private:
-	void RestoreRoundingMode(bool force = false);
-	void ApplyRoundingMode(bool force = false);
-	void UpdateRoundingMode();
-
-	void EatPrefix() { js.EatPrefix(); }
-
-	void FlushAll();
-	void FlushPrefixV();
-
-	u32 GetCompilerPC();
-	void CompileDelaySlot();
-	void EatInstruction(MIPSOpcode op);
-	MIPSOpcode GetOffsetInstruction(int offset);
-
-	// Utility compilation functions
-	void BranchFPFlag(MIPSOpcode op, IRComparison cc, bool likely);
-	void BranchVFPUFlag(MIPSOpcode op, IRComparison cc, bool likely);
-	void BranchRSZeroComp(MIPSOpcode op, IRComparison cc, bool andLink, bool likely);
-	void BranchRSRTComp(MIPSOpcode op, IRComparison cc, bool likely);
-
-	// Utilities to reduce duplicated code
-	void CompShiftImm(MIPSOpcode op, IROp shiftType, int sa);
-	void CompShiftVar(MIPSOpcode op, IROp shiftType, IROp shiftTypeConst);
-
-	void ApplyPrefixST(u8 *vregs, u32 prefix, VectorSize sz);
-	void ApplyPrefixD(const u8 *vregs, VectorSize sz);
-	void GetVectorRegsPrefixS(u8 *regs, VectorSize sz, int vectorReg) {
-		_assert_(js.prefixSFlag & JitState::PREFIX_KNOWN);
-		GetVectorRegs(regs, sz, vectorReg);
-		ApplyPrefixST(regs, js.prefixS, sz);
-	}
-	void GetVectorRegsPrefixT(u8 *regs, VectorSize sz, int vectorReg) {
-		_assert_(js.prefixTFlag & JitState::PREFIX_KNOWN);
-		GetVectorRegs(regs, sz, vectorReg);
-		ApplyPrefixST(regs, js.prefixT, sz);
-	}
-	void GetVectorRegsPrefixD(u8 *regs, VectorSize sz, int vectorReg);
-
-	// Utils
-	void Comp_ITypeMemLR(MIPSOpcode op, bool load);
-
-	// State
-	JitState js;
-	IRWriter ir;
-
-	int dontLogBlocks;
-	int logBlocks;
 };
 
 class IRJit : public JitInterface {
